fix: types 폴더 수정 및 코드 리팩토링
- 코드 리뷰에 따른 코드 구조 수정
- 필요없는 types 삭제
