--- src/types/HistoryCardProps.ts
@@ -1,8 +1,0 @@
-export interface historyCardProps {
-	// historyId: number;
-	// rate: number;
-	url?: string | null;
-	title: string;
-	roadAddress: string;
-	categoryName: string;
-}
